pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.1.0",
    "dspy-ai>=2.4.5",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
    "phonenumbers>=9.0.19",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pywa>=3.5.0",
    "sqlalchemy>=2.0.0",
    "sse-starlette==1.8.2",
    "uvicorn[standard]>=0.24.0",
]
